# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      title:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: request failed
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.SimpleSuccessResponse:
    properties:
      message:
        example: delete success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SwaggerPaginationResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 25
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
  dto.TaskData:
    properties:
      assignee_id:
        example: 2
        type: integer
      created_at:
        type: string
      created_by:
        example: 1
        type: integer
      description:
        example: Mo ta task
        type: string
      id:
        example: 1
        type: integer
      project_id:
        example: 1
        type: integer
      status:
        example: todo
        type: string
      title:
        example: Task 1
        type: string
      updated_at:
        type: string
    type: object
  dto.TaskListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TaskData'
        type: array
      pagination:
        $ref: '#/definitions/dto.SwaggerPaginationResponse'
    type: object
  dto.TaskListSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TaskListPayload'
      message:
        example: get tasks success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.TaskSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TaskData'
      message:
        example: create task success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.UpdateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API cho hệ thống Task Management
  title: Task Management API
  version: "1.0"
paths:
...
  /projects/{id}/tasks:
    get:
      consumes:
      - application/json
      description: Lấy danh sách task theo project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tasks by project
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Tạo task trong project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SimpleSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Lấy chi tiết task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get task detail
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Cập nhật task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - Tasks
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"